TASK: Split responsibilities between NodeService and NodeTypeService
