+++ Classes/Infrastructure/ESCR/NodeService.php
@@ -16,7 +16,6 @@ use Neos\ContentRepository\Core\ContentRepository;
 use Neos\ContentRepository\Core\Feature\SubtreeTagging\Dto\SubtreeTag;
 use Neos\ContentRepository\Core\NodeType\NodeType;
 use Neos\ContentRepository\Core\NodeType\NodeTypeName;
-use Neos\ContentRepository\Core\NodeType\NodeTypeNames;
 use Neos\ContentRepository\Core\Projection\ContentGraph\AbsoluteNodePath;
 use Neos\ContentRepository\Core\Projection\ContentGraph\ContentSubgraphInterface;
 use Neos\ContentRepository\Core\Projection\ContentGraph\Filter\CountChildNodesFilter;
@@ -116,68 +115,6 @@ final class NodeService
         );
 
         return $this->subgraph->findDescendantNodes($rootNode->aggregateId, $filter);
-    }
-
-    public function createNodeTypeFilterFromFilterString(string $filterString): NodeTypeFilter
-    {
-        $nodeTypeManager = $this->contentRepository->getNodeTypeManager();
-        $nodeTypeCriteria = NodeTypeCriteria::fromFilterString($filterString);
-        $explicitlyAllowedNodeTypeNames = $nodeTypeCriteria->explicitlyAllowedNodeTypeNames->toStringArray();
-        $explicitlyDisallowedNodeTypeNames = $nodeTypeCriteria->explicitlyDisallowedNodeTypeNames->toStringArray();
-        $isEmpty =  $nodeTypeCriteria->explicitlyAllowedNodeTypeNames->isEmpty()
-            && $nodeTypeCriteria->explicitlyDisallowedNodeTypeNames->isEmpty();
-
-        $allowedNodeTypeNames = [];
-        foreach ($nodeTypeManager->getNodeTypes(true) as $candidateNodeType) {
-            if ($isEmpty) {
-                $allowedNodeTypeNames[] = $candidateNodeType->name;
-                continue;
-            }
-
-            foreach ($explicitlyDisallowedNodeTypeNames as $disallowedNodeTypeName) {
-                if ($candidateNodeType->isOfType($disallowedNodeTypeName)) {
-                    continue 2;
-                }
-            }
-
-            foreach ($explicitlyAllowedNodeTypeNames as $allowedNodeTypeName) {
-                if ($candidateNodeType->isOfType($allowedNodeTypeName)) {
-                    $allowedNodeTypeNames[] = $candidateNodeType->name;
-                    continue 2;
-                }
-            }
-        }
-
-        return new NodeTypeFilter(
-            nodeTypeCriteria: $nodeTypeCriteria,
-            allowedNodeTypeNames: NodeTypeNames::fromArray($allowedNodeTypeNames)
-        );
-    }
-
-    public function createNodeTypeFilterFromNodeTypeNames(NodeTypeNames $nodeTypeNames): NodeTypeFilter
-    {
-        $nodeTypeManager = $this->contentRepository->getNodeTypeManager();
-        $isEmpty = $nodeTypeNames->isEmpty();
-
-        $allowedNodeTypeNames = [];
-        foreach ($nodeTypeManager->getNodeTypes(true) as $candidateNodeType) {
-            if ($isEmpty) {
-                $allowedNodeTypeNames[] = $candidateNodeType->name;
-                continue;
-            }
-
-            foreach ($nodeTypeNames as $nodeTypeName) {
-                if ($candidateNodeType->isOfType($nodeTypeName)) {
-                    $allowedNodeTypeNames[] = $candidateNodeType->name;
-                    continue 2;
-                }
-            }
-        }
-
-        return new NodeTypeFilter(
-            nodeTypeCriteria: NodeTypeCriteria::createWithAllowedNodeTypeNames($nodeTypeNames),
-            allowedNodeTypeNames: NodeTypeNames::fromArray($allowedNodeTypeNames)
-        );
     }
 
     public function createTreeBuilderForRootNode(
